requirements.txt
flask
tensorflow
numpy
pillow
scipy
gunicorn
